+++ src/implementations/Aave/AaveV3MultiStake.sol
@@ -43,7 +43,7 @@ contract AaveV3MultiStake is StrategyV5 {
    * @param _index Index of the input to stake
    * @param _amount Amount of underlying assets to allocate to `inputs[_index]`
    */
-  function _stake(uint8 _index, uint256 _amount) internal override {
+  function _stake(uint256 _index, uint256 _amount) internal override {
     IAavePool pool = IAavePool(_poolProvider.getPool());
     pool.supply({
       asset: address(inputs[_index]),
@@ -58,7 +58,7 @@ contract AaveV3MultiStake is StrategyV5 {
    * @param _index Index of the input to liquidate
    * @param _amount Amount of underlying assets to recover from liquidating `inputs[_index]`
    */
-  function _unstake(uint8 _index, uint256 _amount) internal override {
+  function _unstake(uint256 _index, uint256 _amount) internal override {
     IAavePool pool = IAavePool(_poolProvider.getPool());
     pool.withdraw({asset: address(inputs[_index]), amount: _amount, to: address(this)});
   }
