+++ src/implementations/Stargate/StargateMultiStake.sol
@@ -69,7 +69,7 @@ contract StargateMultiStake is StrategyV5 {
    */
   function _addLiquiditySingleSide(
     uint256 _amount,
-    uint8 _index
+    uint256 _index
   ) internal returns (uint256 deposited) {
     routers[_index].addLiquidity(poolIds[_index], _amount, address(this));
     return lpTokens[_index].balanceOf(address(this));
@@ -80,7 +80,7 @@ contract StargateMultiStake is StrategyV5 {
    * @param _index Index of the input to stake
    * @param _amount Amount of underlying assets to allocate to `inputs[_index]`
    */
-  function _stake(uint8 _index, uint256 _amount) internal override {
+  function _stake(uint256 _index, uint256 _amount) internal override {
     // deposit+stake
     lpStaker.deposit(stakingIds[_index], _addLiquiditySingleSide(_amount, _index));
   }
@@ -90,7 +90,7 @@ contract StargateMultiStake is StrategyV5 {
    * @param _index Index of the input to liquidate
    * @param _amount Amount of underlying assets to recover from liquidating `inputs[_index]`
    */
-  function _unstake(uint8 _index, uint256 _amount) internal override {
+  function _unstake(uint256 _index, uint256 _amount) internal override {
     // unstake LP
     lpStaker.withdraw(stakingIds[_index], _amount);
     // liquidate LP
